Move read_electrophysiology_anatomical_and_motion_file function to utils.py
