+++ statistics/utils.py
@@ -165,3 +165,69 @@ def read_clinical_file(file_path):
 
     #print(clinical_df)
     return clinical_df
+
+
+def read_electrophysiology_anatomical_and_motion_file(file_path, df_participants):
+    """
+    Read electrophysiology, anatomical, and motion data
+    :param file_path: path to excel file
+    :return anatomical_df: Pandas DataFrame with anatomical data
+    :return motion_df: Pandas DataFrame with motion data
+    :return electrophysiology_df: Pandas DataFrame with electrophysiology data
+    """
+    if os.path.isfile(file_path):
+        print('Reading: {}'.format(file_path))
+        df_all = pd.read_excel(file_path)
+    else:
+        raise FileNotFoundError(f'{file_path} not found')
+
+    # Anatomical data:
+    #   - adapted spinal canal occupation ratio (aSCOR) = spinal cord area/spinal canal area
+    #   - adapted maximum spinal cord compression (aMSCC) = ratio of spinal cord CSA in segment/spinal cord area at C2) - computed only at baseline so far
+    anatomical_df = df_all[['record_id', 'aSCOR_C2', 'aSCOR_C3', 'aSCOR_C4', 'aSCOR_C5', 'aSCOR_C6', 'aSCOR_C7',
+                            'aSCOR_C2_6mth', 'aSCOR_C3_6mth', 'aSCOR_C4_6mth', 'aSCOR_C5_6mth', 'aSCOR_C6_6mth', 'aSCOR_C7_6mth',
+                            'aSCOR_C2_12mth', 'aSCOR_C3_12mth', 'aSCOR_C4_12mth', 'aSCOR_C5_12mth', 'aSCOR_C6_12mth', 'aSCOR_C7_12mth',
+                            'aMSCC_C3toC2', 'aMSCC_C4toC2', 'aMSCC_C5toC2', 'aMSCC_C6toC2', 'aMSCC_C7toC2']]
+
+    # Rename 6mnt and 12mnt columns to 6m and 12m
+    anatomical_df.columns = anatomical_df.columns.str.replace('_6mth', '_6m')
+    anatomical_df.columns = anatomical_df.columns.str.replace('_12mth', '_12m')
+
+    # Insert participant_id column from df_participants to anatomical_df based on record_id
+    anatomical_df = pd.merge(anatomical_df, df_participants[['record_id', 'participant_id']], on='record_id',
+                             how='outer', sort=True)
+    anatomical_df = anatomical_df.set_index(['participant_id'])
+
+    # Segmental motion data:
+    #   - displacement [mm] = area under the curve of the motion plot
+    #   - amplitude [cm/s] = range from maximum positive to maximum negativ velocity values
+    motion_df = df_all[['record_id', 'C2_amp_ax_or_sag', 'C2_disp_ax_or_sag',
+                        'C3_amp_ax_or_sag', 'C3_disp_ax_or_sag',
+                        'C4_amp_ax_or_sag', 'C4_disp_ax_or_sag',
+                        'C5_amp_ax_or_sag', 'C5_disp_ax_or_sag',
+                        'C6_amp_ax_or_sag', 'C6_disp_ax_or_sag',
+                        'C7_amp_ax_or_sag']]
+
+    # Insert participant_id column from df_participants to motion_df based on record_id
+    motion_df = pd.merge(motion_df, df_participants[['record_id', 'participant_id']], on='record_id',
+                         how='outer', sort=True)
+    motion_df = motion_df.set_index(['participant_id'])
+
+    # Electrophysiology data:
+    # - dermatomal SEP (dSEP) with stimulation at C6 and C8 (only few pathologic results)
+    # - dermatomal contact heat evoked potentials (CHEPS) with stimulation at C6, C8 and T4
+    electrophysiology_df = df_all[['record_id', 'dSEP_C6_both_patho_bl', 'dSEP_C8_both_patho_bl', 'CHEPS_C6_patho_bl', 'CHEPS_C8_patho_bl', 'CHEPS_T4_grading_patho_bl',
+                                   'dSEP_C6_both_patho_6mth', 'dSEP_C8_both_patho_6mth', 'CHEPS_C6_patho_6mth', 'CHEPS_C8_patho_6mth', 'CHEPS_T4_patho_6mth',
+                                   'dSEP_C6_both_patho_12mth', 'dSEP_C8_both_patho_12mth', 'CHEPS_C6_patho_12mth', 'CHEPS_C8_patho_12mth', 'CHEPS_T4_patho_12mth',
+                                   'CHEPS_C6_diff_6mth_bl', 'CHEPS_C6_diff_12mth_bl', 'CHEPS_C8_diff_6mth_bl', 'CHEPS_C8_diff_12mth_bl', 'CHEPS_T4_diff_6mth_bl', 'CHEPS_T4_diff_12mth_bl']]
+
+    # Rename 6mnt and 12mnt columns to 6m and 12m
+    electrophysiology_df.columns = electrophysiology_df.columns.str.replace('_6mth', '_6m')
+    electrophysiology_df.columns = electrophysiology_df.columns.str.replace('_12mth', '_12m')
+
+    # Insert participant_id column from df_participants to electrophysiology_df based on record_id
+    electrophysiology_df = pd.merge(electrophysiology_df, df_participants[['record_id', 'participant_id']],
+                                    on='record_id', how='outer', sort=True)
+    electrophysiology_df = electrophysiology_df.set_index(['participant_id'])
+
+    return anatomical_df, motion_df, electrophysiology_df
